requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
colorama==0.4.6
fastapi==0.120.1
grpcio==1.76.0
grpcio-tools==1.76.0
h11==0.16.0
idna==3.11
protobuf==6.33.0
pydantic==2.12.3
pydantic_core==2.41.4
setuptools==80.9.0
sniffio==1.3.1
starlette==0.49.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
